+++ dropwizard-health/src/main/java/io/dropwizard/health/DefaultHealthFactory.java
@@ -9,13 +9,11 @@ import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableList;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
-import io.dropwizard.health.conf.HealthCheckConfiguration;
 import io.dropwizard.health.response.DetailedJsonHealthResponseProviderFactory;
 import io.dropwizard.health.response.HealthResponderFactory;
 import io.dropwizard.health.response.HealthResponseProvider;
 import io.dropwizard.health.response.HealthResponseProviderFactory;
 import io.dropwizard.health.response.ServletHealthResponderFactory;
-import io.dropwizard.health.shutdown.DelayedShutdownHandler;
 import io.dropwizard.jersey.setup.JerseyEnvironment;
 import io.dropwizard.jetty.setup.ServletEnvironment;
 import io.dropwizard.lifecycle.setup.LifecycleEnvironment;
