+++ dropwizard-health/src/main/java/io/dropwizard/health/HealthCheckConfigValidator.java
@@ -1,7 +1,6 @@
 package io.dropwizard.health;
 
 import com.codahale.metrics.health.HealthCheckRegistry;
-import io.dropwizard.health.conf.HealthCheckConfiguration;
 import io.dropwizard.lifecycle.Managed;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -34,35 +33,35 @@ class HealthCheckConfigValidator implements Managed {
     private void validateConfiguration(final List<HealthCheckConfiguration> healthCheckConfigs,
                                        final Set<String> registeredHealthCheckNames) {
         final Set<String> configuredHealthCheckNames = healthCheckConfigs.stream()
-                .map(HealthCheckConfiguration::getName)
-                .collect(Collectors.toSet());
+            .map(HealthCheckConfiguration::getName)
+            .collect(Collectors.toSet());
 
         // find health checks that are registered but do not have a configured schedule
         final Set<String> notConfiguredHealthCheckNames = registeredHealthCheckNames.stream()
-                .filter(healthCheckName -> !configuredHealthCheckNames.contains(healthCheckName))
-                .collect(Collectors.toSet());
+            .filter(healthCheckName -> !configuredHealthCheckNames.contains(healthCheckName))
+            .collect(Collectors.toSet());
 
         if (!notConfiguredHealthCheckNames.isEmpty()) {
             final String healthCheckList = notConfiguredHealthCheckNames.stream()
-                    .map(name -> "  * " + name)
-                    .collect(Collectors.joining("\n"));
+                .map(name -> "  * " + name)
+                .collect(Collectors.joining("\n"));
             LOGGER.info("The following health check(s) were registered, but are not configured with a schedule:\n{}",
-                    healthCheckList);
+                healthCheckList);
         }
 
         // find health checks that are configured with a schedule but are not actually registered
         final Set<String> notRegisteredHealthCheckNames = configuredHealthCheckNames.stream()
-                .filter(healthCheckName -> !registeredHealthCheckNames.contains(healthCheckName))
-                .collect(Collectors.toSet());
+            .filter(healthCheckName -> !registeredHealthCheckNames.contains(healthCheckName))
+            .collect(Collectors.toSet());
 
         if (!notRegisteredHealthCheckNames.isEmpty()) {
             final String healthCheckList = notRegisteredHealthCheckNames.stream()
-                    .map(name -> "  * " + name)
-                    .collect(Collectors.joining("\n"));
+                .map(name -> "  * " + name)
+                .collect(Collectors.joining("\n"));
             LOGGER.error("The following health check(s) are configured with a schedule, but were not registered:\n{}",
-                    healthCheckList);
+                healthCheckList);
             throw new IllegalStateException("The following configured health checks were not registered: "
-                    + notRegisteredHealthCheckNames);
+                + notRegisteredHealthCheckNames);
         }
     }
 }
