+++ dropwizard-health/src/main/java/io/dropwizard/health/HealthCheckType.java
@@ -1,4 +1,4 @@
-package io.dropwizard.health.conf;
+package io.dropwizard.health;
 
 public enum HealthCheckType {
     ALIVE,
