+++ dropwizard-health/src/main/java/io/dropwizard/health/ScheduledHealthCheck.java
@@ -2,13 +2,10 @@ package io.dropwizard.health;
 
 import com.codahale.metrics.Counter;
 import com.codahale.metrics.health.HealthCheck;
-import io.dropwizard.health.conf.HealthCheckType;
-import io.dropwizard.health.conf.Schedule;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.Objects;
-import java.util.Optional;
 
 class ScheduledHealthCheck implements Runnable {
     private static final Logger LOGGER = LoggerFactory.getLogger(ScheduledHealthCheck.class);
@@ -66,7 +63,7 @@ class ScheduledHealthCheck implements Runnable {
 
         HealthCheck.Result result;
         try {
-           result = healthCheck.execute();
+            result = healthCheck.execute();
         } catch (final Exception e) {
             LOGGER.warn("Check for name={} failed exceptionally", name, e);
             result = HealthCheck.Result.unhealthy(e);
@@ -93,12 +90,12 @@ class ScheduledHealthCheck implements Runnable {
         if (!(o instanceof ScheduledHealthCheck)) return false;
         final ScheduledHealthCheck that = (ScheduledHealthCheck) o;
         return critical == that.critical &&
-                Objects.equals(name, that.name) &&
-                Objects.equals(healthCheck, that.healthCheck) &&
-                Objects.equals(schedule, that.schedule) &&
-                Objects.equals(state, that.state) &&
-                Objects.equals(healthyCheckCounter, that.healthyCheckCounter) &&
-                Objects.equals(unhealthyCheckCounter, that.unhealthyCheckCounter);
+            Objects.equals(name, that.name) &&
+            Objects.equals(healthCheck, that.healthCheck) &&
+            Objects.equals(schedule, that.schedule) &&
+            Objects.equals(state, that.state) &&
+            Objects.equals(healthyCheckCounter, that.healthyCheckCounter) &&
+            Objects.equals(unhealthyCheckCounter, that.unhealthyCheckCounter);
     }
 
     @Override
