+++ dropwizard-health/src/main/java/io/dropwizard/health/ShutdownNotifier.java
@@ -1,4 +1,4 @@
-package io.dropwizard.health.shutdown;
+package io.dropwizard.health;
 
 public interface ShutdownNotifier {
     void notifyShutdownStarted() throws Exception;
