+++ dropwizard-health/src/test/java/io/dropwizard/health/HealthCheckConfigValidatorTest.java
@@ -7,7 +7,6 @@ import ch.qos.logback.core.Appender;
 import com.codahale.metrics.health.HealthCheck;
 import com.codahale.metrics.health.HealthCheckRegistry;
 import com.google.common.collect.ImmutableList;
-import io.dropwizard.health.conf.HealthCheckConfiguration;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -35,14 +34,14 @@ public class HealthCheckConfigValidatorTest {
     @BeforeEach
     public void setUp() throws Exception {
         ch.qos.logback.classic.Logger logger = (ch.qos.logback.classic.Logger) LoggerFactory
-                .getLogger(HealthCheckConfigValidator.class);
+            .getLogger(HealthCheckConfigValidator.class);
         logger.addAppender(mockLogAppender);
     }
 
     @AfterEach
     public void tearDown() throws Exception {
         ch.qos.logback.classic.Logger logger = (ch.qos.logback.classic.Logger) LoggerFactory
-                .getLogger(HealthCheckConfigValidator.class);
+            .getLogger(HealthCheckConfigValidator.class);
         logger.detachAppender(mockLogAppender);
         MDC.clear();
     }
@@ -102,11 +101,11 @@ public class HealthCheckConfigValidatorTest {
         LoggingEvent logEvent = captor.getValue();
         assertThat(logEvent.getLevel()).isEqualTo(Level.INFO);
         assertThat(logEvent.getFormattedMessage())
-                .doesNotContain("  * check-1");
+            .doesNotContain("  * check-1");
         assertThat(logEvent.getFormattedMessage())
-                .contains("  * check-2");
+            .contains("  * check-2");
         assertThat(logEvent.getFormattedMessage())
-                .contains("  * check-3");
+            .contains("  * check-3");
     }
 
     @Test
@@ -133,15 +132,15 @@ public class HealthCheckConfigValidatorTest {
             verify(mockLogAppender).doAppend(captor.capture());
             LoggingEvent logEvent = captor.getValue();
             assertThat(logEvent.getLevel())
-                    .isEqualTo(Level.ERROR);
+                .isEqualTo(Level.ERROR);
             assertThat(logEvent.getFormattedMessage())
-                    .doesNotContain("  * check-1");
+                .doesNotContain("  * check-1");
             assertThat(logEvent.getFormattedMessage())
-                    .contains("  * check-3");
+                .contains("  * check-3");
             assertThat(logEvent.getFormattedMessage())
-                    .contains("  * check-3");
+                .contains("  * check-3");
             assertThat(e.getMessage())
-                    .contains("[check-3, check-2]");
+                .contains("[check-3, check-2]");
         }
     }
 }
