+++ dropwizard-health/src/test/java/io/dropwizard/health/HealthCheckManagerTest.java
@@ -5,9 +5,6 @@ import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.health.HealthCheck;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import io.dropwizard.health.conf.HealthCheckConfiguration;
-import io.dropwizard.health.conf.HealthCheckType;
-import io.dropwizard.health.conf.Schedule;
 import io.dropwizard.util.Duration;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -45,7 +42,7 @@ public class HealthCheckManagerTest {
     public void shouldIgnoreUnconfiguredAddedHealthChecks() {
         // given
         final HealthCheckManager manager = new HealthCheckManager(Collections.emptyList(), scheduler,
-                new MetricRegistry(), SHUTDOWN_WAIT, true, Collections.emptyList());
+            new MetricRegistry(), SHUTDOWN_WAIT, true, Collections.emptyList());
 
         // when
         manager.onHealthCheckAdded(NAME, mock(HealthCheck.class));
@@ -62,7 +59,7 @@ public class HealthCheckManagerTest {
         config.setCritical(true);
         config.setSchedule(new Schedule());
         final HealthCheckManager manager = new HealthCheckManager(Collections.singletonList(config), scheduler,
-                new MetricRegistry(), SHUTDOWN_WAIT, true, Collections.emptyList());
+            new MetricRegistry(), SHUTDOWN_WAIT, true, Collections.emptyList());
 
         // when
         manager.onHealthCheckAdded(NAME, mock(HealthCheck.class));
@@ -76,7 +73,7 @@ public class HealthCheckManagerTest {
         // given
         final ScheduledHealthCheck healthCheck = mock(ScheduledHealthCheck.class);
         final HealthCheckManager manager = new HealthCheckManager(Collections.emptyList(), scheduler,
-                new MetricRegistry(), SHUTDOWN_WAIT, true, Collections.emptyList());
+            new MetricRegistry(), SHUTDOWN_WAIT, true, Collections.emptyList());
         manager.setChecks(ImmutableMap.of(NAME, healthCheck));
 
         // when
@@ -90,7 +87,7 @@ public class HealthCheckManagerTest {
     public void shouldDoNothingWhenStateChangesForUnconfiguredHealthCheck() {
         // given
         final HealthCheckManager manager = new HealthCheckManager(Collections.emptyList(), scheduler,
-                new MetricRegistry(), SHUTDOWN_WAIT, true, Collections.emptyList());
+            new MetricRegistry(), SHUTDOWN_WAIT, true, Collections.emptyList());
 
         // when
         manager.onStateChanged(NAME, false);
@@ -108,7 +105,7 @@ public class HealthCheckManagerTest {
         config.setInitialState(false);
         config.setSchedule(new Schedule());
         final HealthCheckManager manager = new HealthCheckManager(Collections.singletonList(config), scheduler,
-                new MetricRegistry(), SHUTDOWN_WAIT, false, Collections.emptyList());
+            new MetricRegistry(), SHUTDOWN_WAIT, false, Collections.emptyList());
         manager.initializeAppHealth();
         final HealthCheck check = mock(HealthCheck.class);
 
@@ -136,7 +133,7 @@ public class HealthCheckManagerTest {
         config.setCritical(true);
         config.setSchedule(new Schedule());
         final HealthCheckManager manager = new HealthCheckManager(Collections.singletonList(config), scheduler,
-                new MetricRegistry(), SHUTDOWN_WAIT, true, Collections.emptyList());
+            new MetricRegistry(), SHUTDOWN_WAIT, true, Collections.emptyList());
         manager.initializeAppHealth();
         final HealthCheck check = mock(HealthCheck.class);
 
@@ -164,7 +161,7 @@ public class HealthCheckManagerTest {
         config.setType(HealthCheckType.ALIVE);
         config.setSchedule(new Schedule());
         final HealthCheckManager manager = new HealthCheckManager(Collections.singletonList(config), scheduler,
-                new MetricRegistry(), SHUTDOWN_WAIT, true, Collections.emptyList());
+            new MetricRegistry(), SHUTDOWN_WAIT, true, Collections.emptyList());
         manager.initializeAppHealth();
         final HealthCheck check = mock(HealthCheck.class);
 
@@ -192,7 +189,7 @@ public class HealthCheckManagerTest {
         config.setCritical(true);
         config.setSchedule(new Schedule());
         final HealthCheckManager manager = new HealthCheckManager(Collections.singletonList(config), scheduler,
-                new MetricRegistry(), SHUTDOWN_WAIT, true, Collections.emptyList());
+            new MetricRegistry(), SHUTDOWN_WAIT, true, Collections.emptyList());
         final HealthCheck check = mock(HealthCheck.class);
 
         // when
@@ -204,29 +201,29 @@ public class HealthCheckManagerTest {
 
         // then
         assertThat(beforeRecovery)
-                .isFalse();
+            .isFalse();
         assertThat(afterRecovery)
-                .isTrue();
+            .isTrue();
         ArgumentCaptor<ScheduledHealthCheck> checkCaptor = ArgumentCaptor.forClass(ScheduledHealthCheck.class);
         ArgumentCaptor<Boolean> healthyCaptor = ArgumentCaptor.forClass(Boolean.class);
         verify(scheduler).scheduleInitial(checkCaptor.capture());
         verify(scheduler, times(2)).schedule(checkCaptor.capture(), healthyCaptor.capture());
         assertThat(checkCaptor.getAllValues().get(0).getName())
-                .isEqualTo(NAME);
+            .isEqualTo(NAME);
         assertThat(checkCaptor.getAllValues().get(0).isCritical())
-                .isTrue();
+            .isTrue();
         assertThat(checkCaptor.getAllValues().get(1).getName())
-                .isEqualTo(NAME);
+            .isEqualTo(NAME);
         assertThat(checkCaptor.getAllValues().get(1).isCritical())
-                .isTrue();
+            .isTrue();
         assertThat(healthyCaptor.getAllValues().get(0))
-                .isFalse();
+            .isFalse();
         assertThat(checkCaptor.getAllValues().get(2).getName())
-                .isEqualTo(NAME);
+            .isEqualTo(NAME);
         assertThat(checkCaptor.getAllValues().get(2).isCritical())
-                .isTrue();
+            .isTrue();
         assertThat(healthyCaptor.getAllValues().get(1))
-                .isTrue();
+            .isTrue();
     }
 
     @Test
@@ -238,7 +235,7 @@ public class HealthCheckManagerTest {
         config.setSchedule(new Schedule());
         final HealthCheck check = mock(HealthCheck.class);
         final HealthCheckManager manager = new HealthCheckManager(Collections.singletonList(config), scheduler,
-                new MetricRegistry(), SHUTDOWN_WAIT, true, Collections.emptyList());
+            new MetricRegistry(), SHUTDOWN_WAIT, true, Collections.emptyList());
         manager.initializeAppHealth();
 
         // when
@@ -264,7 +261,7 @@ public class HealthCheckManagerTest {
         criticalConfig.setSchedule(new Schedule());
         final List<HealthCheckConfiguration> configs = ImmutableList.of(nonCriticalConfig, criticalConfig);
         final HealthCheckManager manager = new HealthCheckManager(configs, scheduler, new MetricRegistry(),
-                SHUTDOWN_WAIT, true, Collections.emptyList());
+            SHUTDOWN_WAIT, true, Collections.emptyList());
         final HealthCheck check = mock(HealthCheck.class);
 
         // when
@@ -278,39 +275,39 @@ public class HealthCheckManagerTest {
 
         // then
         assertThat(beforeRecovery)
-                .isFalse();
+            .isFalse();
         assertThat(afterRecovery)
-                .isFalse();
+            .isFalse();
         ArgumentCaptor<ScheduledHealthCheck> checkCaptor = ArgumentCaptor.forClass(ScheduledHealthCheck.class);
         ArgumentCaptor<Boolean> healthyCaptor = ArgumentCaptor.forClass(Boolean.class);
         verify(scheduler, times(2)).scheduleInitial(checkCaptor.capture());
         verify(scheduler, times(3)).schedule(checkCaptor.capture(), healthyCaptor.capture());
         assertThat(checkCaptor.getAllValues().get(0).getName())
-                .isEqualTo(NAME);
+            .isEqualTo(NAME);
         assertThat(checkCaptor.getAllValues().get(0).isCritical())
-                .isFalse();
+            .isFalse();
         assertThat(checkCaptor.getAllValues().get(1).getName())
-                .isEqualTo(NAME_2);
+            .isEqualTo(NAME_2);
         assertThat(checkCaptor.getAllValues().get(1).isCritical())
-                .isTrue();
+            .isTrue();
         assertThat(checkCaptor.getAllValues().get(2).getName())
-                .isEqualTo(NAME);
+            .isEqualTo(NAME);
         assertThat(checkCaptor.getAllValues().get(2).isCritical())
-                .isFalse();
+            .isFalse();
         assertThat(healthyCaptor.getAllValues().get(0))
-                .isFalse();
+            .isFalse();
         assertThat(checkCaptor.getAllValues().get(3).getName())
-                .isEqualTo(NAME_2);
+            .isEqualTo(NAME_2);
         assertThat(checkCaptor.getAllValues().get(3).isCritical())
-                .isTrue();
+            .isTrue();
         assertThat(healthyCaptor.getAllValues().get(1))
-                .isFalse();
+            .isFalse();
         assertThat(checkCaptor.getAllValues().get(4).getName())
-                .isEqualTo(NAME);
+            .isEqualTo(NAME);
         assertThat(checkCaptor.getAllValues().get(4).isCritical())
-                .isFalse();
+            .isFalse();
         assertThat(healthyCaptor.getAllValues().get(2))
-                .isTrue();
+            .isTrue();
     }
 
     @Test
@@ -343,24 +340,25 @@ public class HealthCheckManagerTest {
             }
 
             @Override
-            public void onStateChanged(String healthCheckName, boolean healthy) {}
+            public void onStateChanged(String healthCheckName, boolean healthy) {
+            }
         };
         final HealthCheckManager manager = new HealthCheckManager(configs, scheduler, metrics, SHUTDOWN_WAIT, true,
             Collections.singleton(countingListener));
 
         final ScheduledHealthCheck check1 = new ScheduledHealthCheck(NAME, READY, nonCriticalConfig.isCritical(), check,
-                schedule, new State(NAME, schedule.getFailureAttempts(), schedule.getSuccessAttempts(), true, manager),
-                metrics.counter(NAME + ".healthy"), metrics.counter(NAME + ".unhealthy"));
+            schedule, new State(NAME, schedule.getFailureAttempts(), schedule.getSuccessAttempts(), true, manager),
+            metrics.counter(NAME + ".healthy"), metrics.counter(NAME + ".unhealthy"));
         final ScheduledHealthCheck check2 = new ScheduledHealthCheck(NAME_2, READY, criticalConfig.isCritical(), check,
-                schedule, new State(NAME, schedule.getFailureAttempts(), schedule.getSuccessAttempts(), true, manager),
-                metrics.counter(NAME_2 + ".healthy"), metrics.counter(NAME_2 + ".unhealthy"));
+            schedule, new State(NAME, schedule.getFailureAttempts(), schedule.getSuccessAttempts(), true, manager),
+            metrics.counter(NAME_2 + ".healthy"), metrics.counter(NAME_2 + ".unhealthy"));
         manager.setChecks(ImmutableMap.of(NAME, check1, NAME_2, check2));
 
         // then
         assertThat(metrics.gauge(manager.getAggregateHealthyName(), null).getValue())
-                .isEqualTo(2L);
+            .isEqualTo(2L);
         assertThat(metrics.gauge(manager.getAggregateUnhealthyName(), null).getValue())
-                .isEqualTo(0L);
+            .isEqualTo(0L);
 
         // when
         when(check.execute()).thenReturn(HealthCheck.Result.unhealthy("because"));
@@ -371,9 +369,9 @@ public class HealthCheckManagerTest {
 
         // then
         assertThat(metrics.gauge(manager.getAggregateHealthyName(), null).getValue())
-                .isEqualTo(1L);
+            .isEqualTo(1L);
         assertThat(metrics.gauge(manager.getAggregateUnhealthyName(), null).getValue())
-                .isEqualTo(1L);
+            .isEqualTo(1L);
         assertThat(unhealthyCounter.get()).isEqualTo(3);
         assertThat(healthyCounter.get()).isEqualTo(0);
     }
@@ -426,9 +424,9 @@ public class HealthCheckManagerTest {
         ArgumentCaptor<ScheduledHealthCheck> checkCaptor = ArgumentCaptor.forClass(ScheduledHealthCheck.class);
         verify(scheduler).scheduleInitial(checkCaptor.capture());
         assertThat(checkCaptor.getValue().getName())
-                .isEqualTo(HealthCheckManagerTest.NAME);
+            .isEqualTo(HealthCheckManagerTest.NAME);
         assertThat(checkCaptor.getValue().isCritical())
-                .isEqualTo(critical);
+            .isEqualTo(critical);
     }
 
     private static class CountingHealthCheck extends HealthCheck {
