+++ dropwizard-health/src/test/java/io/dropwizard/health/config/HealthCheckConfigurationTest.java
@@ -1,8 +1,9 @@
-package io.dropwizard.health.conf;
+package io.dropwizard.health.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.io.Resources;
 import io.dropwizard.configuration.YamlConfigurationFactory;
+import io.dropwizard.health.HealthCheckConfiguration;
 import io.dropwizard.jackson.Jackson;
 import io.dropwizard.jersey.validation.Validators;
 import org.junit.jupiter.api.Test;
@@ -16,7 +17,7 @@ public class HealthCheckConfigurationTest {
     private final ObjectMapper objectMapper = Jackson.newObjectMapper();
     private final Validator validator = Validators.newValidator();
     private final YamlConfigurationFactory<HealthCheckConfiguration> configFactory =
-            new YamlConfigurationFactory<>(HealthCheckConfiguration.class, validator, objectMapper, "dw");
+        new YamlConfigurationFactory<>(HealthCheckConfiguration.class, validator, objectMapper, "dw");
 
     @Test
     public void shouldBuildHealthCheckConfigurationFromYaml() throws Exception {
